requirements.txt
prettytable==0.7.2
six==1.9.0
websocket-client==0.26.0
